test(ci): hold every suite's task names, not just the gates'
The check that a gate names a task some manifest defines sat in the gate
suite's own tests, where it covered twenty-five commands. The topology
builds `deno task` commands in three other places: the two binary
builds, the pattern suites, and the workspace unit suites, which run
each member's own test task and a `browser-test` task beside it. A task
name written wrong in any of them fails the same way a gate does, by
printing the list of tasks that do exist and exiting one, with no
failing test under it to say what broke.

So the check moves up to `tasks/test-topology.test.ts`, which already
loads every suite. It asks each suite for the commands it would build
for all of its units, takes the ones that run a task, and holds the name
to being one Deno can find from where the command runs: defined in that
directory's manifest, or in the root's, which is where Deno looks next.

The message names the suite, the task and the directory, so a run that
fails says which of the four places to look in.
